+++ src/models/Vector.ts
